addressed more issues
