+++ core/src/com/unciv/UncivGame.kt
@@ -17,7 +17,6 @@ import com.unciv.models.tilesets.TileSetCache
 import com.unciv.models.translations.Translations
 import com.unciv.ui.LanguagePickerScreen
 import com.unciv.ui.LoadingScreen
-import com.unciv.ui.audio.CitySoundPlayer
 import com.unciv.ui.audio.GameSounds
 import com.unciv.ui.audio.MusicController
 import com.unciv.ui.audio.MusicMood
@@ -53,14 +52,12 @@ class UncivGame(parameters: UncivGameParameters) : Game() {
     private val customSaveLocationHelper = parameters.customFileLocationHelper
     val platformSpecificHelper = parameters.platformSpecificHelper
     private val audioExceptionHelper = parameters.audioExceptionHelper
-    private val audioExceptionHelper2 = parameters.audioExceptionHelper
 
     var deepLinkedMultiplayerGame: String? = null
     var gameInfo: GameInfo? = null
         private set
     lateinit var settings: GameSettings
     lateinit var musicController: MusicController
-    lateinit var citySoundController: CitySoundPlayer
     lateinit var onlineMultiplayer: OnlineMultiplayer
     lateinit var gameSaver: GameSaver
 
@@ -120,12 +117,6 @@ class UncivGame(parameters: UncivGameParameters) : Game() {
         audioExceptionHelper?.installHooks(
             musicController.getAudioLoopCallback(),
             musicController.getAudioExceptionHandler()
-        )
-
-        citySoundController = CitySoundPlayer()
-        audioExceptionHelper2?.installHooks(
-            citySoundController.getAudioLoopCallback(),
-            citySoundController.getAudioExceptionHandler()
         )
 
         onlineMultiplayer = OnlineMultiplayer()
